chore: Re-structure events model.
